"works"
This revision cleans the code up considerably. It's not as powerful as
it could be.

Particularly, we don't recurse into the generated function trees, we
stay at the top level for swapping.

The interested reader is advised that different generational
strategies are known in the GA literature, this one is an ad-hoc
conglomeration of ones the author had heard about approx. 7.5 years
ago.

Note that depending upon your Lisp system, floating point errors will
be handled differently. Please report floating point bugs to your
maintainers.